6.1.0： 1. 实现了 TopK 算法模板； 2. 规范了命名空间； 3. 去掉了编译器优化； 4. 给 Asan 增加了 coredump 选项配置

+++ src/Base/AbstractBaseContainer.h
@@ -18,7 +18,7 @@ namespace CZ
     {
     public:
         using Rank = uint32_t;
-        using RankPlus = int32_t;
+        using RankPlus = int64_t;
 
         bool empty() const;
